Add a 'gen-prism-grammars' command to 'edb'.
This command generates the prism.js grammars for edgeql and eschema as
well as the synopsis versions of these. The grammars are based on the
keywords as defined in the respective parsers and the introspection of
the schema of an actual EdgeDB database (for purposes of establishing
operators, builtin types and functions).