Reimplement configuration management using new syntax
The current `SET [SYSTEM] CONFIG` syntax is fairly arcane when it comes
to working with non-scalar configuration parameters, such as app ports.

This replaces the current syntax with the following:

1) `CONFIGURE {SYSTEM|SESSION} SET scalar_param := <const-expr>`;
2) `CONFIGURE {SYSTEM|SESSION} RESET scalar_param;`
3) `CONFIGURE SYSTEM INSERT CompositeCfg { prop := ... }`
4) `CONFIGURE SYSTEM RESET CompositeCfg FILTER prop = '...'`

1) and 2) are simple reformulations of the original `SET CONFIG`, but 3)
offers a much cleaner way of specifying a complex composite value by
using the normal `INSERT` syntax.  Likewise, 4) is similar to the normal
`DELETE` statement (although `RESET` is used to emphasize that the
configuration object does not necessarily get deleted, but merely reset
to its value in a previous configuration layer).