graphql: Derive GQL names from fully qualified names.
The type names and base query field names will follow a pattern similar
to what may be expected in EdgeDB. Anything from the 'default' module
uses the shortname, everything else is prefixed with the module name and
uses '__' as the module separator. The 'std' module is still special and
is treated just like 'default', meaning that 'Object' is the name that
gets mapped instead of 'std__Object'.