Remove implicit name-based link and pointer inheritance
Currently concrete links and properties automatically extend abstract
links and properties with the same name in the same module.  This
behavior is historical, since there was originally no way to define
properties on concrete links, so one had to define an abstract link
of the same name with properties and have all concrete links in the
module magically inherit from it.  Today, there is no such restriction,
and, as such, implicit name-based inheritance has no real value, and
is a potential footgun.  Instead, there is now an explicit syntax to
`extend` a link or a property:

    ALTER TYPE Foo { CREATE LINK bar EXTENDING spam -> Bar };