eschema: Convert eschema grammar to SDL closer to DDL.
The language used to defined eschema is currently very similar to DDL
and very different (case-sensitive, whitesapce-sensitive). Which means
that in principle it's a whole new language that needs to be documented
(by us) and learned (by the users). This is a non-trivial burden.
There's also some confusion as to which is primary source of truth - DDL
or eschema?

The new approach is to make the Schema Definition Language (SDL) to be
much more closely related to DDL. The rule of thumb is to make it so
that if you take a CREATE ... statement from DDL and strip out all
the CREATE and SET keywords you should end up with a valid SDL
declaration. The goal of DDL is to specify how to change the DB from
one state to the next, the goal of SDL is to describe the final state.
This is why SDL just mirrors CREATE and SET DDL since we're not
interested in the exact details of the steps.

The grammar has been stripped and moved to the EdgeQL grammar tree, but
the package is still called "eschema" and has a separate codegen and ast.

The command `edb gen-meta-grammars` now only has one valid argument
`edgeql`. However, we still keep two pygments grammars using it
(potentially useful if there's different quirks that need to be
accounted for).

Drop the parentheses from the `EXTENDING` list in both SDL and DDL.

Issue #498.