Remove nested INSERT syntax sugar
Currently the following syntax is supported:

    INSERT Object {
        link: [<TargetType>] {
	    ...
        }
    }

which is equivalent to:

    INSERT Object {
        link := (INSERT TargetType {
            ...
        })
    }

After some discussion, we decided that the shortcut syntax is
problematic, since it exactly matches the syntax in `SELECT` and
`UPDATE`, where it means "match this existing relationship hierarchy".
To avoid possible confusion, especially in the context of `UPDATE`,
remove the shortcut syntax support from `INSERT`.
`foo: [<Type>] [ { <shape> } ]` is now always a selection construct.