Switch storage of "single" links to inline relation attributes
Currently, all links, regardless of cardinality, are represented as a
(source, target) table.  For "single" links this is suboptimal, since an
extra join is required when traversing the link.  The concept of an
"inline" storage is already used for `__type__`, so this change merely
extends it for all non-multi links and fixes the bugs and omissions this
had exposed.

The downside of this change is that changing the cardinality of a link
is now less trivial and requires moving the data during the migration.