Implement initial support for user authentication
This adds generic authentication framework via the new `cfg::Auth`
configuration object.  The two initial authentication modes are `Trust`
and `SCRAM` (SCRAM-SHA-256).  The default is `SCRAM`, which means that a
valid role password is required to connect.   Since new instances
obviously don't have that configured, add a way to connect to a special
"admin" UNIX socket, which is protected by OS permissions, and which
does not enforce internal authentication.

To make things easier, the development server (which is now started with
`edb server`) has `Trust` auth configured by default.