Add support for managing roles via a CLI
This is the first command of a series of CLI commands that allow
basic manipulation of the EdgeDB instance and its schema using command
line syntax.  The basic notion is that `edgedb {create|alter|drop} ...`
follow a familiar DDL syntax, minus the unnecessary punctuation.

`edgedb {create,alter,drop} role` is chosen to be implemented as the first
such command to make it easier to set up a role from the command line
during initial EdgeDB setup.