Raise an exception on invalid input to mean(), var(), and stddev()
`math::mean()` is not defined on empty input set, `var()` and `stddev()`
are not defined on sets with cardinality less than 2.  Rather then
silently returning an empty set (which is a result of an underlying SQL
function returning NULL), raise an explicit exception.

This commit also adds a missing `InvalidValueError` mapping that fixes
the incorrect error being raised on parameter validation errors.  This
commit also wires in support for passing the query source context on
runtime errors, if available.