Add initial support for enumerated types
Enumerated types are a special class of a scalar type that are defined
as a list of explicit string values using the following syntax:

    CREATE SCALAR TYPE <name> EXTENDING enum<'val1'[,...'valN']>

Over the wire, and in text serialization, the enum values are transferred
as strings, but they sort according to the order of enum value definitions,
and every enum type is distinct, i.e. there is no implicit cast between
them.

Currently, there is no way to alter the composition of enumerated type
after it was created.