edgeql: Enforce the use of parentheses in "ON (...)" constructs.
The syntactical construct "ON Expr { ..." is ambiguous. The "{" can be
the start of a shape or a sub-command block. Thus the parentheses are
required for disambiguating the grammar.

This also means that "INDEX ... ON (...)" should use parentheses.
Currently the reason is consistency, but in principle the INDEX commands
should support nested blocks for setting attributes.