Fix deficiencies in the enum implementation
It appears that subtyping enums in Postgres strips away the ability to
apply any operators on them, so stop using domains of enums, prohibit
enumerated type subclassing, defaults, and constraints on enumerated
types.  This also adds support for enums in SDL and adds handling for
the most common enum data validation error.