Prevent schema objects with live references from being dropped
Use schema reference tracking to prohibit dropping schema objects that
have at least one live schema object referencing it.  This mechanism
obviously relies on the completeness of the schema reference tracking,
so this protection is not complete yet.  Most notably, expressions are
not considered as sources of references yet, so it's still possible to
corrupt the schema by dropping an object used in a view, for example.
Nonetheless, this is an important step toward ensuring that the schema
is always consistent.