Use the data_dir as runstate_dir when bootstrapping
When bootstrapping a new EdgeDB instance it is often necessary to avoid
using the main runstate dir due to lack of permissions, possibility of
conflict with another running instance, etc.  Since the server is not
really available during bootstrap it makes sense to use the target
data_dir as runstate_dir.