docs: Clean up SDL and DDL docs
Rather than duplicating the synopsys description refer to DDL for most
things and make a separate mention of the nested sub-declarations (liked
to SDL declarations).

In DDL be more explicit in the synopsis about which subcommands are
avaialable, but refer to `CREATE...` blocks when describing `ALTER...`
blocks to avoid duplication.

Add an "Administration" section under "EdgeQL". Move "Databases" from
DDL to the admin section. Add commands for roles.

Also there are few minor tweaks and fixes to exposed fields.