Switch to SERIALIZABLE as the default transaction isolation level
Continuing the theme of safety and consistency by default, switch the
default transaction isolation level to `SERIALIZABLE`.  Remove the
`READ COMMITTED` level as it does not provide any tangible benefits over
the higher isolation levels.  `REPEATABLE READ` remains as an option for
scenarios that are pathological for the current implementation of
`SERIALIZABLE` in Postgres, as well as for use in hot-standbys.