Rename "attributes" to "annotations"
The rename makes the feature pretty much self-explanatory.  The reason
for the original naming is historical: all schema fields were
attributes.  At some point, it was decided that we'll have system
attributes (called fields) like "required" or "readonly" and
user-definable text attributes ("attributes").

Now it's obvious that the "attribute" name is simply confusing and
"annotation" is a better alternative.