graphql: Update the type-reflection logic.
Tuples are not reflected.
Arrays are reflected as long as they are not of json or bytes.
Bytes are not reflected.
If a scalar is not even reconglized as being from the list of known
scalars (such as name changed in std, or a new scalar added to std) an
exception is raised.