Allow expressions to return generic types in certain scenarios
Currently, a bare empty set constructor (`{}`) or a bare array
literal (`[]`) are illegal without an explicit type cast.  There are
cases, however, where the target type is now known, but the expression
still needs to be processed, such as the `initial_value` attribute of
a generic function.  Generic functions, such as `std::count` also don't
_need_ the input to be of a concrete type.

Thus, define `{}` as returning `anytype` and `[]` as returning
`array<anytype>`.  Operator and function resolution would naturally
prohibit using these, unless explicitly allowed.  It is also illegal for
a query to return a value of a generic type (this applies to computables
as well).