Fix a couple of protocol issues found when documenting it
The startup sequence has been shortened, the connection parameters are
no longer sent as a separate message, and are instead sent as part of
the initial `ClientHandshake` message.

The `ServerKeyData` message now contains a sequence of 32 bytes instead
of a 32-bit integer.

This is a breaking change, but no bump to the protocol version, since
we're in alpha.