edgeql: Allow str to be assigned to an enum or be used as default.
Enums themselves are not strings, but the only way to construct one is
by casting a str into it. In case of assignment, there's already no
ambiguity as to what the specific enum type is needed and the
convenience of not having to write the full cast every time is
desirable. Similarly, when a default value is specified for an enum
property, there's no need to have an explicit cast.

Assignment casts are now enforced at the EdgeQL->IR level.