Fix issues with implicit shapes
If requested, the compiler generates an implicit computable
`__tid__` into every object shape that goes to output.  Currently, if
the expression did not have an explicit shape, the `__tid__` would get
added to the *base* type, which is obviously wrong.

The second issue here is that nested shapes did not trigger the
derivation of the pointer in the parent shape, which, although currently
ignored by the compiler, is also incorrect.