Replace C ext row_getattr_by_num with open-coded implementation
Originally, row_getattr_by_num() was added as a workaround for older
Postgres versions that had issues with dealing with anonymous row types.
Version 11 has those fixed, and so the workaround is no longer necessary
and can be replaced by

    (SELECT attr FROM unnest(ARRAY[record]) AS <tupdesc>)

This approach is also faster, so remove the C version altogether.