Stop using default attr for computables
Currently, computable poiners store their expression in the 'default'
attr.  The original reasoning was that setting defaults for computable
pointers has no sense anyway, and defaults are used as a 'computable' in
mutation assignments.  However, computable expressions and default
expressions have different requirements, most importantly, the
`__source__` magic is prohibited in defaults, but is perfectly legal in
a computable.  For this reason, store the computable expression in a
dedicated attr.  This makes the current 'computable' attr unnecessary,
so the net attr count is unchanged.