Fix rename DDL
The renaming of references (e.g links and properties) is now checked for
sanity with respect to inheritance.  A reference cannot be renamed if
it's inherited from one or more ancestors that aren't *also* being
renamed.

Additionally, renaming of schema objects that are used in some
expression elsewhere is prohibited, since we don't yet have a way to
automatically recompile all dependent expressions.