Fix issues with effect visibility of DML subqueries
When a DML statement is used as an expression it is imperative that the
effects of a DML statement are visible when querying data paths
originating from the subquery expression.  Since all Postgres CTEs
execute in the same snapshot, the effects must be correctly propagated
by overlaying the DML CTE result onto the base object/pointer rels.  The
mechanism to do so already existed in the form of
`Environment.rel_overlays`, but the implementation lacked in
specificity and had a number of bugs which caused either the failure to
find a correct overlay or set it up correctly.

The fixes also necessitated the disbanding of `dbobj` by moving all
context-dependent utilities to `relctx` and context-agnostic utilities
ot `astutils`.  This structure makes more sense and gives the former
`dbobj` utilities full access to the context and not just the
`Environment` part.