Fix buggy behavior of LIMIT inside a view
LIMIT and OFFSET currently have rather hacky implementation that breaks
down completely if the statement is used to define a view.  There is no
particular reason to juggle the scope to implement the necessary
`limit(SET OF subj, SET OF limit)` semantics, as a simple AST reorg with
an implicit `SELECT` wrapper in the left argument is entirely
sufficient.