Fix a bunch of bugs and omissions in migration DDL deparsing
Current implementation of `GET MIGRATION` is rusty and poorly tested,
and this brings things to a much saner state.  There is a new
rudimentaty test suite here intended to check that DDL obtained from
`GET MIGRATION` produces the exact same result as `COMMIT MIGRATION` on
the same migration.  This is a prerequisite of making sure that schema
dumps are produced correctly.