Make nested views properly introspectable
Currently, only the first level of a view is introspectable, which makes
nested views useless in contexts where proper introspection is
necessary, such as in GraphQL reflection.

Making schema views fully introspectable requires two changes: the names
of nested derived views must be unique and properly formed (i.e not use
the generic derived mangling), the target refs in of the derived
pointers in the view must be kept.

Also, `DROP VIEW` is currently broken when used on a view with nesting,
since it leaves behind the derived types created for nested views.  That
is also fixed by this commit.