Place restrictions on the use of DML statements
The behavior of DML CTEs in Postgres dictates that they: a) must be a
top-level CTE; b) always execute to completion. These restrictions must
be respected by the semantics of EdgeQL DML. The top-level CTE
requirement can only be satisfied if the use of correlated vars inside a
DML statement is prohibited, i.e:

    SELECT (Foo, (INSERT Bar { foo := Foo }));

The execute-to-completion requirement can be satisfied by prohibiting
the use of DML inside a conditional construct, like `??` or `IF`.

Closes: #741