Remove SELECT-like clauses from FOR statement
Currently, the `FOR` statement accepts the `FILTER`, `ORDER BY`,
`OFFSET` and `LIMIT` clauses that apply to the *result* of the `UNION`.
In essense, `FOR x IN <expr> UNION <expr> FILTER <flt>` is equivalent to
`SELECT <implicit_alias> := (FOR x IN <expr> UNION <expr>) FILTER <flt>`

The problem is that the above transformation is arguably not intuitive,
since it is possible to *also* assume that the `FILTER` clause etc.
apply to the expression in the `UNION` clause.  The current
implementation is also incorrect, since it doesn't hide the iterator
variable from the followup clauses, and doing that would also be an
exception from all other statements.

For those reasons, remove the clause sugar support from the `FOR`
statement until we figure out an intuitive way of reintroducing them.