Hoist iterator scopes when compiling computables in nested shapes
Nested shape annotations are usually specified syntactically nested in a
subquery, e.g:

    WITH MODULE test
    SELECT User {
        select_deck := (
            SELECT _ := (
                FOR letter IN {'I', 'B'}
                UNION (
                    SELECT User.deck {
                        name,
                        letter := letter
                    }
                    FILTER .name[0] = letter
                )
            )
            ORDER BY _.name
        )
    } FILTER .name = 'Alice';

In the example above, the objects selected in the `select_deck`
computable will have the `name` and the `letter` properties, the
latter is itself a computable.   The canonical IR form for
such nested shape structure is actually as follows:

    WITH MODULE test
    SELECT User {
        select_deck := (
            SELECT _ := (
                FOR letter IN {'I', 'B'}
                UNION (
                    SELECT User.deck
                    FILTER .name[0] = letter
                )
            )
            ORDER BY _.name
        ) {
            name,
            letter := <letter>
        }
    } FILTER .name = 'Alice';

Note that the nested shape has been pulled out.  However, the
computable `letter` references the iterator variable, and, if that
symbol isn't pulled out to the scope where the nested shape is now,
the cardinality of the `letter` computable would be wrong.