Restore the upstream behavior of injecting type OIDs into row data
Although EdgeDB itself ignores backend OIDs in the row and array data,
filtering them out is unnecessary work, and elision on the Postgres
side requires a patch, so just live with a couple of "reserved" words
in the protocol messages.

Alas, we also now have to fix the query argument bytestream supplied
by the client by injecting the correct OID into every array datum.
That, obviously, requires EdgeDB to track the OID for every type that
may be an array element, so this introduces the `Type.backend_id`
schema field.