Track type coverage of the codebase with tests
Tracking per function.  Also allows for tracking per line within functions but
that's not that important and way more noisy as it would require changing tests
on almost every change to every file, very annoying.  This is also why we're
not using Mypy's builtin per-line coverage reporting here.

Partially-typed function signatures are counted as untyped.  This is to
differentiate between old "documentation-style" annotations and actual
annotations covered by Mypy.

Those tests are set up to also fail when the actual percentage is *higher* than
the expected percentage.  This is so that we can raise the expectation in time
and keep the bar.

This is part of #693.