edgeql: Store original index expression in a special annotation.
Add a new ANNOTATION `schema::system` for storing details used by
EdgeDB. Users should not change this annotation directly.

Record the original INDEX expression text in the `system` annotation for
that index. This way it is possible to use that to display a more
human-friendly expression during introspection as well. It is also used
to identify the INDEX itself in SDL and DDL.

Update the syntax so that index annotations are allowed. Also, disallow
indexes as top-level statements in DDL or SDL. They must be nested
inside a type or a link statement.

Fixes: #758