Fix some inconsistencies in ir.ast
`ir.ast.Set.expr` is defined as `ir.ast.Base`, which is wrong, since it
must always be an instance of `ir.ast.Expr`.  Fix this, and the
relevant typing annotations in the EdgeQL frontend.

This also fixes a couple of remaining typing errors in
`edb.edgeql.compiler`, so enable mypy checks on it.