Rewrite "import typing" into "from typing import *" (#783)
There's a script based on Black that does this in edb/tools/rewrite_typing.py.

The script checks for shadowing of typing by names local to the file in which case it skips such a file. The script also skips files that say "import typing # NoQA" so we can force keeping nominal imports for some files where it's needed (like the flake8 plugin).

This change also removes remaining string annotations and adjusts the typing coverage test to reject them in the future.