Add the new DESCRIBE command
Currently, there is no way for an EdgeDB client to obtain an SDL or DDL
representation of a given part of a schema or a part of a schema, except
by using introspection queries and implementing client-side code
generation. Since the facilities to produce correct schema dumps already
exist in the server, it makes a lot of sense to expose the functionality
as commands:

    DESCRIBE { <schema-object> [, ...] | SCHEMA }
        [AS {DDL|SDL|TEXT}] [VERBOSE] ;

`DDL`, `SDL` and `TEXT` determine the output format, SDL being the default.

`<schema-object>` is a valid reference to a schema object: a (possibly
qualified) name, or an object type followed by a name. If the object is
global, then the type is required.

If the entire schema is to be described, the sole `SCHEMA` keyword is used.

Examples:

    # Return the description of the `User` object type as SDL
    DESCRIBE TYPE User AS SDL;

    # Return the description of the `orders` module as DDL.
    DESCRIBE MODULE orders AS DDL;

    # Return the description of the entire schema as DDL.
    DESCRIBE SCHEMA AS DDL;

    # Return the annotated description of the `User` type with verbosity
    DESCRIBE TYPE User AS TEXT VERBOSE;

Closes: #790