Fix problems with test teardown when running `edb test`
This is yet another *attempt* to convince our multiprocessed
unittest to actually do final test teardown properly.

Both `unittest` and `multiprocessing` are at fault here, with
multiple issues to address.

Problem ONE: same `tearDownClass` called multiple times.  This is
a result of misunderstanding of what `unittest.TestSuite` actually
is.  Our current teardown fix iterates over the registered
`TestSuite` instances and calls `_tearDownPreviousClass` on them.
This is wrong, because `TestSuite` is essentially just a bunch
of methods that mutate a *single result instance*, so they're really
basically static methods of `TestResult`, and so must be called
**once**, and not per `TestSuite` instance as is now.

Problem TWO: `multiprocessing.Pool` worker finalization does not
work.  When running with `-jN`, where `N` is greater than one,
we use the standard process pool to run tests in parallel.
Currently, the teardown portion is implemented via
`multiprocessing.util.Finalize`, which is a complicated proxy
for `atexit()`.  Using `atexit()` for finalization is a fraught
undertaking, since there's no guarantee if any object would be
alive at this point (hence the `Finalize` hacks with its
`weakref` use).

A sane approach is to actually provide a counterpart of
the `initialize` argument to the Pool constructor to ensure
correct worker cleanup and call it just after the worker
loop concludes.  The only way to do this currently is to
monkeypatch `multiprocessing.pool.worker()`.

But this is not all.  Since `multiprocessing` does not actually
care about correct worker finalization, it basically just
calls `Process.terminate()` on them as soon as `Pool.terminate()`
is called, without trying to be nice and `join()` them properly.
Hence another monkeypatch to convince it to do so.