Postpone interpretation of schema refs in pointer definitions
Currently, we attempt to do too much at the AST->DDL IR stage for links
and properties: we validate the target type and the computable expression
in the context where we might not have all the necessary types yet.
Most importantly, this breaks tight loop cases like this:

    create type Category {
        create link parent -> Category
    };

As a general rule, any assumption about the schema state at the AST->IR
stage of DDL processing is invalid, and only the checks that don't rely
on the schema can be safely made.  Schema-dependent checks must be made
during the DDL canonicalization pass.