Stop using C-extension functions for parsing date and time.
Use regexp to validate the presense of time zone spec in the format
string for `to_datetime` and `to_local_datetime` (and other local
converters).

Also, use a special trick to validate that time zone is
actually present in the input as well (because `to_timestamp` simply
doesn't care). The trick involves temporarily changing the server time
zone settings and seeing if that affects the conversion.

Issue #740.