Reimplement floor division operator in pure SQL.
Implement floor division and its companion modular division in pure SQL.
The basic principle is that integer division result is always rounded
down (as opposed to rounded to 0). This has the desirable property that
modular division always yields positive results for positive modulus,
while giving negative results for negative modulus.

Mathematically they are equivalent, but practically modular division
that always produces the results between 0 and the modulus is useful,
since these are often used as offsets (in, say, arrays) where 1 and -1
(for mod 2) are not equivalent.

Fixes #753.