tests: Bring type coverage under the "cqa" label.
Add the "cqa" prefix to type coverage tests as they also
represent code QA. The label is meant to be a shorthand for all the
tests that need to be run regardless of the change. It is a way to quickly
re-test a minor change or update without re-running the whole test suite
while avoiding pesky errors in CI that appear to do code style rather
than correctness.

Flake8 and mypy tests are already marked with "cqa" prefix.