edgeql: Remove SELECT-like clauses from FOR.
The `FOR` statement now only has the basic form of:

    FOR x IN {iterexpr} UNION expr;

All the `FILTER`, `ORDER BY`, `LIMIT` and `OFFSET` clauses have been
removed due to the potential ambiguity of their meaning (as evidenced by
even some tests with incorrect semantics).

If those clauses are necessary, then the correct procedure is to wrap
the whole `FOR ...` expression into a `SELECT`.

When rewriting the tests to move the SELECT-like clauses some potential
scoping issues were dicovered.

Fixes #743.