Add @profile() decorator
It lives under `edb.tools.profiling`. This decorator has a few unique
properties:

- it allows for reusing the same profiler instance across calls, making it
  feasible to use also for functions called millions of times

- it uses temporary files in ways that enable using it across processes

- then provides an `aggregate()` function to gather multiple saved profile
  files into one result that can be observed

There is a simple command-line client for the `aggregate()` functionality
invoked by `python -m edb.tools.profiling`.

Note: by default on macOS temporary files are created in a deterministic
directory which is *not* /tmp/.  If you don't pass `dir=` to the @profile()
decorator, also don't pass a directory to the aggregate function (or the
command-line client) and the function will find the directory for you.

There are tests.