edgeql: Update backtick quoting.
Add the ability to use a backtick inside a backtick-quoted identifier by
using a double-backtick "``". Primarily the double-backtick is chosen
over backslash-quoting "\`" because the latter could be misleading. No
other backslash-escape sequence is interpreted inside backticks so
using, say "\n", in an identifier would be accepted, but not result in a
newline character.

Update the error message for identifiers starting with "@" to be more
specific.

Fixes: #632.