Disable runtime checks in production by default
Runtime type checks in edb.common.ast and edb.common.checked are taking up to
15% of the compilation cost based on profiling done as part of #694.

This change disables runtime type checks  by default unless an
EDGEDB_DEBUG_TYPECHECK=1 flag is passed in the environment.  Note: this is also
true for tests with the exception of tests of the runtime typechecking
capability itself.

In the case of edb.common.checked, disabling type checks cannot unfortunately
be done by replacing the data structures with builtin untyped versions due to
picklability.  Instead, we're shadowing the type checking classmethods from the
abstract classes with identity functions on the concrete types.

Benchmarks to follow.