Add basic `\d` command to REPL.
Add `\d NAME` command to REPL. It is a shortcut for `DESCRIBE OBJECT`
where the `NAME` is plain text name of the concept that needs to be
described. No backtick name quoting is needed, but it can be used.

For example, the following are equivalent:

    \d O bje`ct
    \d `O bje``ct`

Largely the backticks are optional so that when keywords are used as
names no quoting is necessary. For example, describing a user-created
type `Commit` could be done like this:

    \d Commit
    \d default::Commit
    \d default::`Commit`

Issue: #179