Mask the F811 flake8 error on both Python 3.7 & 3.8
In Python 3.8 the position of a decorated function in AST points
to the line where the "def" keyword is; in Python 3.7 it points
to the first decorator.  This change breaks the behavior of
flake8 w.r.t. how to mask F811.

Relevant flake8 issue: https://gitlab.com/pycqa/flake8/issues/583