Fix comparison of collections of distinct (but compatible) types
The current recursive resolution of operators on collections incorrectly
assumes that a mere presence of a valid operator for the elements of the
collection is enough, and that would be OK, if operators on collections
were actually implemented in terms of the element-wise application of
the correct operator of the element types.  Unfortunately, current
Postgres does not support comparison of arrays or rows of distinct
types, even if a corresponding operator for the element types exists.
Fortunately, since the generic recursive collection operator is defined
as polymorphic on all parameters, the resulting call binding will force
a cast on the arguments to bring them to the same type.