edgeql: Update the SDL syntax.
The `MIGRATION` DDL commands now must use an unqualified name for the
migration. The idea is that migrations are global objects and represent
the entire schema.

The SDL used in the `CREATE MIGRATION` command now must follow these
rules:
0) It ignores the default module from `SET MODULE` command. Anything
from the `std` module may be referenced without being fully-qualified.
1) The SDL describes the entire user-created schema.
2) A `module <name> { ... }` may wrap SDL declarations belonging to the
same module. Inside the module block top-level names MUST be unqualified
and are always assumend to belong to the module named by the block. All
other names (extending, targets, etc.) may be fully-qualified.
Unqualified names are interpreted as if the module named in the block
declaration is currently the default module.
3) Module blocks cannot be nested.
4) Any declaration appearing outside of a `module` block MUST use
fully-qualified names at top-level. The module portion of the
fully-qualified name will be used to determine which module the declared
entity belongs to.

Issue #907