Prohibit NaN as a std::decimal value
NaN in PostgreSQL numeric is a weird thing.  First of all, none of the
special IEEE 754 values are supported (i.e. there is no infinity).  Zero
division always raises an exception, so `0/0` is not a thing either.
Similarly, other "undefined" operations, such as `log(-1)` are also an
error.

Given this apparent lack of usefulness, and a significant number of
potential pitfalls (comparison weirdness, virality etc), it seems better
to just ban the use of NaNs in `std::decimal`.