Prohibit implicit comparisons between anynumeric and anyfloat
Comparing decimals and bigints with floats without an explicit cast is
now prohibited.  This is mostly for consistency with our general stance
on the decimal/float separation, but also because current operators
don't work for collection comparisons, which require a common implicit
cast, and there is none between floats and decimals.