Fix nested INSERT into a single link with properties
Using a nested INSERT to populate a `single` link with properties
currently crashes the compiler due to an old kludge that was trying to
do an ad-hoc tuple-serialization of a nested insert.  This hack is no
longer necessary, and removing it fixes the crash.  This fix unmasked
another issue: the sub-INSERT was visited twice due to how link
properties and object properties are routed in DML, so a basic memo and
a guard are added to prevent having duplicate DML CTEs.