Change the scoping logic for tuples and tuple dereference
Currently, expressions in the form `A.b.c` are handled in the same way
regardless of whether it is an object pointer dereference or a tuple
element dereference.  Since tuple elements are *always* singletons, the
common prefix ambiguity restriction we have for object paths is not very
useful, so remove it.  In practical terms this means that the following
expression is now valid:

    WITH
        MODULE test,
        letter := {'A', 'B'},
        tup := (
            letter,
            (
                SELECT User
                FILTER .name[0] = letter
            )
        )
    SELECT tup.1 {
        l := tup.0,
    };