Fix an issue with inherited constraint name.
There are two ways of inheriting a constraint:
1) Two types are created, one extending the other, then the
constraint is added to the base type.
2) The base type is created with a constraint and then it's extended by
another type.

These two ways should end up with the same constraints on the extending
type, in particular the internal generated name of the concrete
constraint should be the same in both cases. Names that don't match can
cause problems in migrations or simply fail to be properly processed.