Simplify signature of ir.typeutils.ptrref_from_ptrcls
The `ptrref_from_ptrcls()` function is currently taking a bunch of
arguments that it can actually derive from the `ptrcls` argument.
So, in ord[er to make the function easier to cache, remove the extraneous
parameters.

This makes `ptrref_from_ptrcls()` somewhat more expensive, since it has
to do more work, but this will be compensated by a followup change to
cache its results.