Optimize single link type indirections when possible
When the type filter expression is applied on a type that is a
*supertype* of a type in the filter expression, we have the opportunity
to optimize the resulting query by joining the target type var directly.
This optimization is currently only applied to multi links, but there's
no real reason for this restriction, so extend the optimization to
single links as well.