Make unions produced by reverse link navigation universal
Currently, a reverse link path expression (e.g `User.<owner`) produces a
regular union type of all possible types that have an `owner` link to
`User`.  While this may seem fine at a glance, this behavior actually
leads to fragile queries due to the "action at a distance" effect.
The meaning of the expression might change unintentionally via the
addition of new types (not all of which are directly controlled by the
user).

For this reason, make it so that reverse link navigation produces a
univesral union, i.e the union of all possible types, which is
`std::Object`.  Then, to access further links or properties (other than
the ones defined on `std::Object`), an explicit type filter operator
(`[IS ..]`) must be used, so instead of `User.<owner.number` one must to
write `User.<owner[IS Issue].number`.

While working on this I also realized that there is a subtle issue
related to how link properties are treated in the context of link
unions.  First of all, this commit actually makes paths like
`User.<owner[IS Issue]@note` work.  Notably, the implementation here
*does not* attempt to generate link unions based on inheritance
resolution of link source, so in `Foo.<link[IS Bar]` the link `Bar.link`
must actually exist.  Whether to change this so that a union of `link`
links of all subtypes of `Bar` is generated is left for further
discussion and work.

Fixes: #969