Fix issues with rebasing and constraints.
It is important that alternative DDL paths that lead to the same
conceptual state of the database should in fact produce identical
schemas. This includes all the names and bases for the affected
concepts. It is especially important for concrete constraints that were
created by different DDL paths as they have automatically-generated
names which are then used in determining the behavior of the next
migration or other schema alterations.