Rename VIEW to ALIAS
Currently, we have the `CREATE VIEW` command.  It creates a special
object in the schema that is, essentially, an expression of a certain
type.  The problem with the name `VIEW` is that this is not really a
"view" in a common database sense.  Instead, `CREATE VIEW` is
essentially equivalent to a schema-level `WITH` block declaration.

This

    CREATE VIEW MyUser := (SELECT User { compute_prop := <expr> });
    SELECT MyUser { compute_prop };

is semantically equivalent to this:

    WITH MyUser := (SELECT User { compute_prop := <expr> })
    SELECT MyUser { compute_prop };

We call declarations in the `WITH` block "aliases".  Specifically, you
can declare a *module alias* (`WITH foo AS MODULE bar`), and you can
declare an *expression alias* as in the example above. Given this, it's
evident that the current use of the keyword `VIEW` is not an optimal
choice, and that `ALIAS` should be used instead. This is especially
reinforced by the existing syntax that declares *session-level* aliases:

    SET ALIAS foo AS MODULE bar;

So, an equivalent session-level expression alias would be declared by:

    SET ALIAS MyUser := (SELECT User { compute_prop := <expr> });

This makes `SET ALIAS` and `WITH` perfectly symmetrical.

Final point: we will implement "proper" views at some point. Unlike
expression aliases, they would define an entirely distinct type (as
opposed to merely augmenting, and masquerading as, an existing type).

Fixes: #989